Covered populations for all states at state, county, census tract level
This is an optional dataset that includes all of the covered variables, including rural populations and incarcerated populations. A guide is included. Caveat -- this dataset does not relay how incarcerated populations were calculated; I tried to compare the numbers in the dataset with various census variables that I had identified related to incarcerated populations, and there was no pattern I could discern.